requirements.txt
Django==3.2
python-dotenv==0.21.0
Pillow==9.4.0
geopy==2.3.0
sorl-thumbnail==12.9.0
requests==2.28.2
folium==0.14.0
django-filter==22.1
django-debug-toolbar==3.8.1